[chore] Change module name and add requirements
